Ruff format (#104)
* applied ruff format to the whole code base
* enforce whitespace and ruff as part of the CI.
(ruff check is not enforced yet. this will come in the next PR)

+++ src/chemcoord/cartesian_coordinates/asymmetric_unit_cartesian_class.py
@@ -23,16 +23,19 @@ class AsymmetricUnitCartesian(Cartesian):
         Returns:
             Cartesian: A new cartesian instance.
         """
-        coords = ['x', 'y', 'z']
-        eq_sets = self._metadata['eq']['eq_sets']
-        sym_ops = self._metadata['eq']['sym_ops']
-        frame = pd.DataFrame(index=[i for v in eq_sets.values() for i in v],
-                             columns=['atom', 'x', 'y', 'z'], dtype='f8')
-        frame['atom'] = pd.Series(
-            {i: self.loc[k, 'atom'] for k, v in eq_sets.items() for i in v})
+        coords = ["x", "y", "z"]
+        eq_sets = self._metadata["eq"]["eq_sets"]
+        sym_ops = self._metadata["eq"]["sym_ops"]
+        frame = pd.DataFrame(
+            index=[i for v in eq_sets.values() for i in v],
+            columns=["atom", "x", "y", "z"],
+            dtype="f8",
+        )
+        frame["atom"] = pd.Series(
+            {i: self.loc[k, "atom"] for k, v in eq_sets.items() for i in v}
+        )
         frame.loc[self.index, coords] = self.loc[:, coords]
         for i in eq_sets:
             for j in eq_sets[i]:
-                frame.loc[j, coords] = np.dot(sym_ops[i][j],
-                                              frame.loc[i, coords])
+                frame.loc[j, coords] = np.dot(sym_ops[i][j], frame.loc[i, coords])
         return Cartesian(frame)
